отрефакторил CounterPlugin с использованием стримов

+++ main-module/src/main/java/ru/digitalhabbits/homework1/service/FileEngine.java
@@ -1,5 +1,7 @@
 package ru.digitalhabbits.homework1.service;
 
+import org.slf4j.Logger;
+
 import javax.annotation.Nonnull;
 import java.io.File;
 import java.io.IOException;
@@ -9,16 +11,20 @@ import java.nio.channels.FileLock;
 import java.nio.channels.OverlappingFileLockException;
 
 import static java.util.Arrays.stream;
+import static org.slf4j.LoggerFactory.getLogger;
 
 public class FileEngine {
+    private final Logger logger = getLogger(FileEngine.class);
     private static final String RESULT_FILE_PATTERN = "results-%s.txt";
     private static final String RESULT_DIR = "results";
     private static final String RESULT_EXT = "txt";
 
     public boolean writeToFile(@Nonnull String text, @Nonnull String pluginName) {
         // TODO: NotImplemented
-        String fileName = String.format(RESULT_FILE_PATTERN, pluginName);
-        try{
+        String resultDirPath = System.getProperty("user.dir") + RESULT_DIR;
+        String fileName = resultDirPath + "\\" + String.format(RESULT_FILE_PATTERN, pluginName);
+        logger.info("'{}' results are in '{}'", pluginName, fileName);
+        try {
             RandomAccessFile stream = new RandomAccessFile(fileName, "rw");
             FileChannel channel = stream.getChannel();
 
@@ -44,7 +50,10 @@ public class FileEngine {
     public void cleanResultDir() {
         final String currentDir = System.getProperty("user.dir");
         final File resultDir = new File(currentDir + "/" + RESULT_DIR);
-        stream(resultDir.list((dir, name) -> name.endsWith(RESULT_EXT)))
-                .forEach(fileName -> new File(resultDir + "/" + fileName).delete());
+        logger.info("Path to result directory is '{}'",resultDir.getName());
+        if (resultDir.list() != null) {
+            stream(resultDir.list((dir, name) -> name.endsWith(RESULT_EXT)))
+                    .forEach(fileName -> new File(resultDir + "/" + fileName).delete());
+        }
     }
 }
